requirements.txt

numpy
scikit-learn==1.2.1
matplotlib==3.7.2
pyyaml
torch
torchvision